[AMSDK-11143] - Assertion fail on misread of persistence in TestPersistence helper method

+++ code/consentedge/src/androidTest/java/com/adobe/marketing/mobile/TestPersistenceHelper.java
@@ -19,6 +19,8 @@ import com.adobe.marketing.mobile.consent.ConsentTestConstants;
 
 import java.util.ArrayList;
 
+import static org.junit.Assert.fail;
+
 /**
  * Helper class to update and remove persisted data to extension concerned with testing Consents.
  */
@@ -31,29 +33,56 @@ public class TestPersistenceHelper {
 
     /**
      * Helper method to update the {@link SharedPreferences} data.
+     *
      * @param datastore the name of the datastore to be updated
-     * @param key the persisted data key that has to be updated
-     * @param value the new value
+     * @param key       the persisted data key that has to be updated
+     * @param value     the new value
      */
     public static void updatePersistence(final String datastore, final String key, final String value) {
         final Application application = TestHelper.defaultApplication;
+        if (application == null) {
+            fail("Unable to updatePersistence by TestPersistenceHelper. Application is null, fast failing the test case.");
+        }
+
         final Context context = application.getApplicationContext();
-        SharedPreferences sharedPreferences = context.getSharedPreferences(datastore, Context.MODE_PRIVATE);;
+        if (context == null) {
+            fail("Unable to updatePersistence by TestPersistenceHelper. Context is null, fast failing the test case.");
+        }
+
+        SharedPreferences sharedPreferences = context.getSharedPreferences(datastore, Context.MODE_PRIVATE);
+
+        if (sharedPreferences == null) {
+            fail("Unable to updatePersistence by TestPersistenceHelper. sharedPreferences is null, fast failing the test case.");
+        }
+
         SharedPreferences.Editor editor = sharedPreferences.edit();
-        editor.putString(key,value);
+        editor.putString(key, value);
         editor.apply();
     }
 
     /**
      * Reads the requested persisted data from datastore.
+     *
      * @param datastore the name of the datastore to be read
-     * @param key that needs to be read
-     * @return the persisted data in {@code String}
+     * @param key       the key that needs to be read
+     * @return {@link String} value of persisted data. Null if data is not found in {@link SharedPreferences}
      */
     public static String readPersistedData(final String datastore, final String key) {
         final Application application = TestHelper.defaultApplication;
+        if (application == null) {
+            fail("Unable to readPersistedData by TestPersistenceHelper. Application is null, fast failing the test case.");
+        }
+
         final Context context = application.getApplicationContext();
-        SharedPreferences sharedPreferences = context.getSharedPreferences(datastore, Context.MODE_PRIVATE);;
+        if (context == null) {
+            fail("Unable to readPersistedData by TestPersistenceHelper. Context is null, fast failing the test case.");
+        }
+
+        SharedPreferences sharedPreferences = context.getSharedPreferences(datastore, Context.MODE_PRIVATE);
+        if (sharedPreferences == null) {
+            fail("Unable to readPersistedData by TestPersistenceHelper. sharedPreferences is null, fast failing the test case.");
+        }
+
         return sharedPreferences.getString(key, null);
     }
 
@@ -61,10 +90,23 @@ public class TestPersistenceHelper {
      * Clears the Configuration and Consent extension's persisted data
      */
     public static void resetKnownPersistence() {
+
         final Application application = TestHelper.defaultApplication;
+        if (application == null) {
+            fail("Unable to resetPersistence by TestPersistenceHelper. Application is null, fast failing the test case.");
+        }
+
         final Context context = application.getApplicationContext();
+        if (context == null) {
+            fail("Unable to resetPersistence by TestPersistenceHelper. Context is null, fast failing the test case.");
+        }
+
         for (String eachDatastore : knownDatastoreName) {
-            SharedPreferences sharedPreferences = context.getSharedPreferences(eachDatastore, Context.MODE_PRIVATE);;
+            SharedPreferences sharedPreferences = context.getSharedPreferences(eachDatastore, Context.MODE_PRIVATE);
+            if (sharedPreferences == null) {
+                fail("Unable to resetPersistence by TestPersistenceHelper. sharedPreferences is null, fast failing the test case.");
+            }
+
             SharedPreferences.Editor editor = sharedPreferences.edit();
             editor.clear();
             editor.apply();
